+++ test/mock_hardware.dart
@@ -86,26 +86,10 @@ class MockHardware implements RpiGPIO {
   }
 
   @override
-  int gpioNum(int pinNum) =>
-      [17, 18, 27, 22, 23, 24, 25, 4, 2, 3, 8, 7, 10, 9, 11, 14, 15][pinNum];
-
-  @override
   void initInterrupts(SendPort port) {
     if (interruptEventPort != null) throw 'interrupts already initialized';
     interruptEventPort = port;
   }
-
-  @override
-  int physPinToGpio(int pinNum) => [
-        -1, //                                      0
-        -1, -1, 2, -1, 3, -1, 4, 14, -1, 15, //     1 - 10
-        17, 18, 27, -1, 22, 23, -1, 24, 10, -1, // 11 - 20
-        9, 25, 11, 8, -1, 7, 0, 1, 5, -1, //       21 - 30
-        6, 12, 13, -1, 19, 16, 26, 20, -1, 21, //  31 - 40
-        -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, // 41 - 50
-        28, 29, 30, 31, -1, -1, -1, -1, -1, -1, // 51 - 60
-        -1, -1, -1, -1, -1, -1, -1, -1, //         60+
-      ][pinNum];
 
   void reset() {
     values = <int>[0, 0, 0, 0, null];
@@ -141,6 +125,12 @@ class MockHardware implements RpiGPIO {
   void setTrigger(int pinNum, Trigger trigger) {
     if (interruptEventPort == null) throw 'must call initInterrupts';
     interruptMap[pinNum] = trigger;
+  }
+
+  @override
+  String description(int pinNum) {
+    // Simulate sparse mock where this is not implemented
+    throw new NoSuchMethodError(this, new Symbol('description'), [pinNum], {});
   }
 }